changing order's view

+++ Application/ViewModels/Product/ProductViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,8 +10,11 @@ namespace Application.ViewModels.Product
     public class ProductViewModel
     {
         public int Id { get; set; }
+
+
         public string ProductName { get; set; }
 
-        public int Price { get; set; }
+
+        public decimal Price { get; set; }
     }
 }
